# openapi-onboarding.json
{"openapi":"3.1.0","info":{"title":"Momna E17 Onboarding Flow API","version":"1.0.0","description":"Canonical client onboarding transport over Momna Universal Flow Engine; bearer identity is server-derived and raw canonical context is never returned."},"paths":{"/api/v1/flow-instances":{"post":{"operationId":"startOrResumeFlow","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Authoritative flow state"},"400":{"$ref":"#/components/responses/Error"},"401":{"$ref":"#/components/responses/Error"},"409":{"$ref":"#/components/responses/Error"},"503":{"$ref":"#/components/responses/Error"}}}},"/api/v1/flow-instances/{instanceId}":{"get":{"operationId":"getFlow","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Authoritative flow state"},"401":{"$ref":"#/components/responses/Error"},"403":{"$ref":"#/components/responses/Error"},"404":{"$ref":"#/components/responses/Error"}}}},"/api/v1/flow-instances/{instanceId}/answers/{fieldId}":{"put":{"operationId":"putFlowAnswer","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Recalculated flow state"},"400":{"$ref":"#/components/responses/Error"},"401":{"$ref":"#/components/responses/Error"},"403":{"$ref":"#/components/responses/Error"},"409":{"$ref":"#/components/responses/Error"}}}},"/api/v1/flow-instances/{instanceId}/actions/{action}":{"post":{"operationId":"flowAction","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Recalculated flow state"},"400":{"$ref":"#/components/responses/Error"},"409":{"$ref":"#/components/responses/Error"}}}},"/api/v1/flow-instances/{instanceId}/complete":{"post":{"operationId":"completeFlow","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Safe completion projection"},"409":{"$ref":"#/components/responses/Error"}}}},"/api/v1/flow-instances/{instanceId}/result":{"get":{"operationId":"recoverFlowResult","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Safe result recovery without raw answers"}}}},"/api/v1/flow-instances/{instanceId}/router/confirm":{"post":{"operationId":"confirmRouter","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Lifecycle confirmed and period onboarding launched"},"409":{"$ref":"#/components/responses/Error"}}}}},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}},"responses":{"Error":{"description":"Stable redacted error envelope","content":{"application/json":{"schema":{"type":"object","required":["code","message","requestId"],"properties":{"code":{"type":"string"},"message":{"type":"string"},"requestId":{"type":"string"}},"additionalProperties":false}}}}}}}